move the cluster.py to folder /app/cluster
